vcytcongreso2016/requirements.txt
Django==1.8.6
django-smart-selects==1.1.1
Pillow==3.0.0
wheel==0.24.0
